+++ example/lib/src/hls_example_page.dart
@@ -4,10 +4,16 @@ import 'package:vlc_player/vlc_player.dart';
 import 'player_example_view.dart';
 
 class HlsExamplePage extends StatefulWidget {
-  const HlsExamplePage({super.key, this.showPlayer = true, this.source});
+  const HlsExamplePage({
+    super.key,
+    this.showPlayer = true,
+    this.source,
+    this.playerOptions = const <String>[],
+  });
 
   final bool showPlayer;
   final Uri? source;
+  final List<String> playerOptions;
 
   @override
   State<HlsExamplePage> createState() => _HlsExamplePageState();
@@ -18,6 +24,7 @@ class _HlsExamplePageState extends State<HlsExamplePage> {
     source:
         widget.source ??
         Uri.parse('https://test-streams.mux.dev/x36xhzz/x36xhzz.m3u8'),
+    options: widget.playerOptions,
   );
 
   @override